# [stdlib] Make `_memmem` function use Span