Make `_memmem` function use Span.

This is a split-off from PR #3548

### Benchmarks

Intel i7-7700HQ

  | old value (ms) | new value (ms) | markdown percentage
-- | -- | -- | --
bench_string_find_single[10] | 6.48115018378378 | 6.03484829081633 | 0.0689
bench_string_find_multiple[10] | 17.9806469444444 | 12.2977545454545 | 0.3161
bench_string_find_single[30] | 15.1792885822785 | 15.3715400655738 | -0.0127
bench_string_find_multiple[30] | 39.4704182068966 | 26.4202728913043 | 0.3306
bench_string_find_single[50] | 11.30529816 | 11.8535879 | -0.0485
bench_string_find_multiple[50] | 29.2628455208333 | 22.2432352727273 | 0.2399
bench_string_find_single[100] | 6.85366217142857 | 7.07481515 | -0.0323
bench_string_find_multiple[100] | 17.4309957391304 | 14.7496196790123 | 0.1538
bench_string_find_single[1000] | 21.1665719821429 | 34.3364576756757 | -0.6222
bench_string_find_multiple[1000] | 66.363384 | 55.7598185 | 0.1598
bench_string_find_single[10000] | 168.445427666667 | 173.191917857143 | -0.0282
bench_string_find_multiple[10000] | 508.217139666667 | 338.433613 | 0.3341
  |   | Average | 0.0716