@gryznar 2025-03-28
Here also it will be good to know impact on performance ;)