@martinvuyk 2025-08-21
@JoeLoser forgot I was benchmarking on smaller lengths originally. I just added the `for _ in range(10**6 // length)` limit where it was previously just running a million iterations for very long strings as well.